Add dummy testing function to notify service
Testing out the testing framework jest. Created a dummy js function to test.
That function always return true. To run the test, navigate to the folder
with a CLI and run 'npm install --save-dev jest' the first time you do it.
Then run 'npm test'. This runs jest, which looks for the following files
in the current folder or its subfolders:
Files with .js suffix in __tests__ folders.
Files with .test.js suffix.
Files with .spec.js suffix.
And runs them and shows you the result.