Test pusher beam functionality
Trying out the pusher beam notification service.
To test it out for yourself. Go to the localhost:PORT,
Click on "Register with Pusher beams", and "Subscribe to Test message".
This will subscribe the user to a type of notifications (interest).
Then run "node send_test_notification.js" to send the notification.