Implement first version of scheduling module using stubs
This commit adds a scheduling module that uses stubs that print to console
to simulate the behavior of sending Notification to particular user at
scheduled times. The module exports a function startScheduler that takes an
optional argument frequencyInSeconds, which specifies the
frequency of the job in seconds, defaulting to 300 (5 minutes).

The startScheduler function creates a ToadScheduler object and
two jobs: a task that runs the getUsersAndInterests and
sendNotifications stubs, and a SimpleIntervalJob that runs the
task at the specified frequency. Before starting this
SimpleIntervalJob, it waits to the beginning of the next 5 minute
interval using either a busy wait or a non-blocking wait (the
latter is currently broken and commented out).